fastapi_app/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
jinja2
python-dotenv
passlib[bcrypt]
itsdangerous
alembic
starlette-sessions
pydantic-settings
pytest
httpx
python-multipart
pytest-asyncio
